Added security context to all services working with DB

+++ quiz/src/main/java/com/klopsi/answer/AnswerService.java
@@ -1,11 +1,15 @@
 package com.klopsi.answer;
 
 import com.klopsi.answer.model.Answer;
+import com.klopsi.exercise.model.Exercise;
 import lombok.NoArgsConstructor;
 import javax.enterprise.context.ApplicationScoped;
+import javax.inject.Inject;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
+import javax.servlet.http.HttpServletRequest;
 import javax.transaction.Transactional;
+import java.security.AccessControlException;
 import java.util.List;
 
 @ApplicationScoped
@@ -14,25 +18,47 @@ public class AnswerService {
 	@PersistenceContext
 	private EntityManager em;
 
+	@Inject
+	private HttpServletRequest securityContext;
+
 	public List<Answer> findAllAnswers() {
-		return em.createNamedQuery(Answer.Queries.FIND_ALL, Answer.class).getResultList();
+		if(securityContext.isUserInRole("USER")){
+			return em.createNamedQuery(Answer.Queries.FIND_ALL, Answer.class).getResultList();
+		}
+		else {
+			throw new AccessControlException("Access denied");
+		}
 	}
 
 	public synchronized Answer findAnswer(int id) {
-		return em.find(Answer.class, id);
+		if(securityContext.isUserInRole("USER")){
+			return em.find(Answer.class, id);
+		}
+		else {
+			throw new AccessControlException("Access denied");
+		}
 	}
 
 	@Transactional
 	public void removeAnswer(Answer answer) {
-		em.remove(em.merge(answer));
+		if(securityContext.isUserInRole("ADMIN")){
+			em.remove(em.merge(answer));
+		}
+		else {
+			throw new AccessControlException("Access denied");
+		}
 	}
 
 	@Transactional
 	public synchronized void saveAnswer(Answer answer){
-		if (answer.getId() == null) {
-			em.persist(answer);
-		} else {
-			em.merge(answer);
+		if(securityContext.isUserInRole("USER")){
+			if (answer.getId() == null) {
+				em.persist(answer);
+			} else {
+				em.merge(answer);
+			}		}
+		else {
+			throw new AccessControlException("Access denied");
 		}
 	}
 }
